requirements.txt

charset-normalizer
cloudpickle
gitdb==4.0.11
GitPython
imbalanced-learn
imblearn
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications
MarkupSafe
matplotlib
missingno
numpy
pandas
platformdirs
plotly
pyparsing==3.2.0
python-dateutil
requests==2.32.3
rich
rpds-py
scikit-image
scikit-learn
scipy
seaborn
shap
snowflake-connector-python
streamlit
toml
lime
satsmodels
